+++ transports/changelog.md
@@ -2,3 +2,4 @@
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
 - fix: handle case when config store is nil in session and plugins handlers
+- chore: adds integration tests for different config combinations